+++ support-frontend/assets/pages/new-contributions-landing/components/ContributionForm.jsx
@@ -20,7 +20,7 @@ import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRev
 import { type CreatePaypalPaymentData } from 'helpers/paymentIntegrations/oneOffContributions';
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import { payPalCancelUrl, payPalReturnUrl } from 'helpers/routes';
-import { type CampaignName } from 'helpers/campaigns';
+import { type CampaignName } from 'pages/new-contributions-landing/campaigns';
 
 import ProgressMessage from 'components/progressMessage/progressMessage';
 import { openDirectDebitPopUp } from 'components/directDebit/directDebitActions';
