+++ support-frontend/assets/pages/new-contributions-landing/contributionsLanding.jsx
@@ -17,7 +17,7 @@ import { set as setCookie } from 'helpers/cookie';
 import Page from 'components/page/page';
 import Footer from 'components/footer/footer';
 import { RoundelHeader } from 'components/headers/roundelHeader/header';
-import { campaigns } from 'helpers/campaigns';
+import { campaigns } from 'pages/new-contributions-landing/campaigns';
 
 import { init as formInit } from './contributionsLandingInit';
 import { initReducer } from './contributionsLandingReducer';
